WIP - UNSTABLE - 8 hours later full contract replacements
I feel we have enough material to show how we could change things piece by piece in a stable fashion.

The competitor replacements on the other hand also show the fact that it's like setting off a bomb in the entire application. 
Basic Immutability practices could have solved this issue.